refactor: make students on clubService
